requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
openai
dateparser